refactor: move constants.ts
